+++ apps/auto-email-sender-new/src/index.ts
@@ -4,15 +4,61 @@ import { setupMatchRoles } from 'shared/src/services/setupMatchRoles'
 import { rolesRenderer } from './roles-renderer'
 import { subsToQueue } from './subs-to-queue'
 import { composeEmail } from './email-composer'
-import { createRabbitMqChannel } from 'shared'
+import path from 'path'
+import os from 'os'
+import axios from 'axios'
+import { delay } from 'shared/utils/utils'
 
 const execPromise = promisify(exec)
+const numCores = os.availableParallelism()
+
+console.log(numCores)
+
+const runCommand = promisify(exec)
+
+const checkMatchRolesUp = async (
+  url: string,
+  maxAttempts = 10,
+  interval = 5000
+) => {
+  let attempt = 0
+  while (attempt < maxAttempts) {
+    try {
+      const response = await axios.get(url)
+      if (response.status === 200) {
+        console.log('match_roles is up.')
+        return
+      }
+    } catch (error) {
+      console.error(`Error checking match_roles: ${error}`)
+    }
+    attempt++
+    console.log(
+      `Attempt ${attempt} failed. Retrying in ${interval / 1000} seconds...`
+    )
+    await delay(interval)
+  }
+  throw new Error(`match_roles did not start after ${maxAttempts} attempts.`)
+}
 
 export const init = async () => {
+  const matchRolesPath = path.resolve(
+    'apps/auto-email-sender-new/src/match_roles'
+  )
   try {
-    // console.log('Starting match_roles');
-    // await setupMatchRoles();
-    // console.log('match_roles completed.');
+    console.log('Starting match_roles')
+    runCommand(
+      'bash -c "source matchenv/bin/activate && cd src && uvicorn --host 0.0.0.0 main:app"',
+      {
+        cwd: matchRolesPath,
+      }
+    )
+
+    console.log('Waiting for match_roles to be up')
+    await checkMatchRolesUp(
+      'http://127.0.0.1:8000/best_role?skills=25%2C40%2C450&languages=English%2CPortuguese&n=2'
+    )
+    console.log('match_roles is up and running.')
 
     console.log('Starting subsToQueue...')
     await subsToQueue()
@@ -47,9 +93,13 @@ export const init = async () => {
 }
 
 const runParallelAssignRoles = async () => {
-  const tasks = Array(4)
+  const tasks = Array(numCores)
     .fill(null)
-    .map(() => execPromise('ts-node src/utils/assignRolesWorker.ts'))
+    .map(() =>
+      execPromise(
+        'ts-node apps/auto-email-sender-new/src/utils/assignRolesWorker.ts'
+      )
+    )
   try {
     await Promise.all(tasks)
   } catch (err: any) {
@@ -60,7 +110,7 @@ const runParallelAssignRoles = async () => {
 }
 
 const runParallelEmailPreRender = async () => {
-  const tasks = Array(4)
+  const tasks = Array(numCores)
     .fill(null)
     .map(() => execPromise('ts-node src/utils/emailPreRenderWorker.ts'))
   try {
